WIP #62 Moved test classes into namespaces matching service assembly
