+++ src/DNI.Services/Podcast/PodcastStream.cs
@@ -1,10 +1,7 @@
 ﻿using System.Collections.Generic;
 
-using RestSharp.Deserializers;
-
 namespace DNI.Services.Podcast {
     public class PodcastStream {
-        [DeserializeAs(Name = "items")]
         public List<PodcastShow> Shows { get; set; } = new List<PodcastShow>();
     }
 }
