+++ test/DNI.Services.Tests/Podcast/FiresidePodcastServiceIntegrationTests.cs
@@ -4,10 +4,9 @@ using AutoFixture;
 using AutoFixture.AutoMoq;
 
 using DNI.Options;
-using DNI.Services.Vodcast;
+using DNI.Services.Podcast;
 using DNI.Testing;
 
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
@@ -18,45 +17,37 @@ using RestSharp;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace DNI.Services.Tests {
+namespace DNI.Services.Tests.Podcast {
     [Trait(TraitConstants.TraitTestType, TraitConstants.TraitTestTypeIntegration)]
-    public class YouTubeVodcastServiceIntegrationTests {
+    public class FiresidePodcastServiceIntegrationTests {
         private readonly ITestOutputHelper _output;
         private readonly IFixture _fixture = new Fixture().Customize(new AutoMoqCustomization {ConfigureMembers = true});
         private readonly IOptions<GeneralOptions> _generalOptions;
-        private readonly IOptions<YouTubeOptions> _youTubeOptions;
-        private readonly Mock<ILogger<YouTubeVodcastService>> _loggerMock;
+        private readonly Mock<ILogger<FiresidePodcastService>> _loggerMock;
 
-        private readonly IConfiguration config;
-
-        public YouTubeVodcastServiceIntegrationTests(ITestOutputHelper output) {
+        public FiresidePodcastServiceIntegrationTests(ITestOutputHelper output) {
             _output = output;
 
-            _loggerMock = Mock.Get(_fixture.Create<ILogger<YouTubeVodcastService>>());
+            _loggerMock = Mock.Get(_fixture.Create<ILogger<FiresidePodcastService>>());
 
             // Set up default / valid options
             var generalOptions = _fixture.Create<GeneralOptions>();
             _generalOptions = Microsoft.Extensions.Options.Options.Create(generalOptions);
-            _generalOptions.Value.VodcastServiceBaseUri = "https://www.googleapis.com/youtube/v3";
-
-            config = TestHelpers.GetConfigFromFile();
-            _youTubeOptions = config.CreateOptions<YouTubeOptions>("YouTubeOptions");
+            _generalOptions.Value.PodcastServiceResourceUri = "dnistream/rss";
+            _generalOptions.Value.PodcastServiceBaseUri = "https://feeds.fireside.fm";
         }
 
-        /// <summary>
-        ///     The below test will only succeed if an APIKey is provided
-        /// </summary>
-        /// <returns></returns>
         [Fact]
         public async Task GetAllAsync_ReturnsDataFromRemoteUri() {
             // Arrange
             var restClient = new RestClient();
-            var service = new YouTubeVodcastService(restClient, _generalOptions, _youTubeOptions, _loggerMock.Object);
+            var service = new FiresidePodcastService(restClient, _generalOptions, _loggerMock.Object);
 
             // Act
             var r = await service.GetAllAsync();
 
             // Assert
+            Assert.NotNull(r);
             Assert.NotNull(r.Shows);
             Assert.True(r.Shows.Count > 0);
         }
