+++ test/DNI.Services.Tests/Podcast/FiresidePodcastServiceUnitTests.cs
@@ -19,7 +19,7 @@ using RestSharp;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace DNI.Services.Tests {
+namespace DNI.Services.Tests.Podcast {
     [Trait(TraitConstants.TraitTestType, TraitConstants.TraitTestTypeUnit)]
     public class FiresidePodcastServiceUnitTests {
         private readonly ITestOutputHelper _output;
