+++ test/DNI.Services.Tests/Vodcast/YouTubeVodcastServiceIntegrationTests.cs
@@ -4,9 +4,10 @@ using AutoFixture;
 using AutoFixture.AutoMoq;
 
 using DNI.Options;
-using DNI.Services.Podcast;
+using DNI.Services.Vodcast;
 using DNI.Testing;
 
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
@@ -17,37 +18,45 @@ using RestSharp;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace DNI.Services.Tests {
+namespace DNI.Services.Tests.Vodcast {
     [Trait(TraitConstants.TraitTestType, TraitConstants.TraitTestTypeIntegration)]
-    public class FiresidePodcastServiceIntegrationTests {
+    public class YouTubeVodcastServiceIntegrationTests {
         private readonly ITestOutputHelper _output;
         private readonly IFixture _fixture = new Fixture().Customize(new AutoMoqCustomization {ConfigureMembers = true});
         private readonly IOptions<GeneralOptions> _generalOptions;
-        private readonly Mock<ILogger<FiresidePodcastService>> _loggerMock;
+        private readonly IOptions<YouTubeOptions> _youTubeOptions;
+        private readonly Mock<ILogger<YouTubeVodcastService>> _loggerMock;
 
-        public FiresidePodcastServiceIntegrationTests(ITestOutputHelper output) {
+        private readonly IConfiguration config;
+
+        public YouTubeVodcastServiceIntegrationTests(ITestOutputHelper output) {
             _output = output;
 
-            _loggerMock = Mock.Get(_fixture.Create<ILogger<FiresidePodcastService>>());
+            _loggerMock = Mock.Get(_fixture.Create<ILogger<YouTubeVodcastService>>());
 
             // Set up default / valid options
             var generalOptions = _fixture.Create<GeneralOptions>();
             _generalOptions = Microsoft.Extensions.Options.Options.Create(generalOptions);
-            _generalOptions.Value.PodcastServiceResourceUri = "dnistream/rss";
-            _generalOptions.Value.PodcastServiceBaseUri = "https://feeds.fireside.fm";
+            _generalOptions.Value.VodcastServiceBaseUri = "https://www.googleapis.com/youtube/v3";
+
+            config = TestHelpers.GetConfigFromFile();
+            _youTubeOptions = config.CreateOptions<YouTubeOptions>("YouTubeOptions");
         }
 
+        /// <summary>
+        ///     The below test will only succeed if an APIKey is provided
+        /// </summary>
+        /// <returns></returns>
         [Fact]
         public async Task GetAllAsync_ReturnsDataFromRemoteUri() {
             // Arrange
             var restClient = new RestClient();
-            var service = new FiresidePodcastService(restClient, _generalOptions, _loggerMock.Object);
+            var service = new YouTubeVodcastService(restClient, _generalOptions, _youTubeOptions, _loggerMock.Object);
 
             // Act
             var r = await service.GetAllAsync();
 
             // Assert
-            Assert.NotNull(r);
             Assert.NotNull(r.Shows);
             Assert.True(r.Shows.Count > 0);
         }
